+++ api/db.py
@@ -1,4 +1,4 @@
 from sqlalchemy import create_engine
 
-db = create_engine('sqlite:///_database.db')
+db = create_engine('sqlite:///database.db')
 db.connect()
