+++ main/twooter-app/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import Sidebar from './Sidebar';
-import Feed from './Feed';
-import Widgets from './Widgets';
-import "./App.css";
+import Sidebar from './Components/Sidebar';
+import Feed from './Components/Feed';
+import Widgets from './Components/Widgets';
+import "./styles/App.css";
 
 function App() {
   return (
