+++ main/twooter-app/src/Components/Feed.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Feed.css';
+import '../styles/Feed.css';
 import TweetBox from './TweetBox';
 import Post from './Post';
