+++ main/twooter-app/src/Components/Sidebar.js
@@ -1,5 +1,5 @@
 import React from "react";
-import './Sidebar.css'
+import '../styles/Sidebar.css'
 import SidebarOption from './SidebarOption';
 
 // Icons
