+++ main/twooter-app/src/Components/Widgets.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import './Widgets.css';
+import '../styles/Widgets.css';
 import SearchIcon from "@material-ui/icons/Search"
 import Widget from './Widget';
